# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Faculty Library Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /loans:
    get:
      operationId: listAllLoans
      tags: [loans]
      summary: List all loans in the system (Librarian/Admin)
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: bookId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, RETURNED, OVERDUE]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Paginated loans
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedLoans"
...
  /loans/borrow:
    post:
      operationId: borrowBook
      tags: [loans]
      summary: Borrow a book (atomic transaction)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BorrowBookBody"
      responses:
        "201":
          description: Loan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "400":
          description: Cannot borrow (quota, duplicate, unavailable)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /loans/{loanId}/return:
    post:
      operationId: returnLoan
      tags: [loans]
      summary: Process a return (Librarian/Admin)
      parameters:
        - name: loanId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Loan returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{loanId}/renew:
    post:
      operationId: renewLoan
      tags: [loans]
      summary: Renew a loan (borrower only, max 2 renewals)
      parameters:
        - name: loanId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Loan renewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "400":
          description: Cannot renew
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Users ──────────────────────────────────────────────────────────────────
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (Admin only)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID (Admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user role or active status (Admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── Dashboard ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [STUDENT, FACULTY, LIBRARIAN, ADMIN]
        department:
          type: string
        phone:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, department, isActive, createdAt]
    UpdateUserBody:
      type: object
      properties:
        role:
          type: string
          enum: [STUDENT, FACULTY, LIBRARIAN, ADMIN]
        isActive:
          type: boolean
    PaginatedUsers:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        page:
          type: integer
        pageSize:
          type: integer
        totalRecords:
          type: integer
        totalPages:
          type: integer
      required: [data, page, pageSize, totalRecords, totalPages]

    # ─── Book Schemas ───────────────────────────────────────────────────────
    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        isbn:
          type: string
        publisher:
          type: string
        publicationYear:
          type: integer
        edition:
          type: string
          nullable: true
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        format:
          type: string
          enum: [PHYSICAL, DIGITAL]
        totalCopies:
          type: integer
        availableCopies:
          type: integer
        shelfLocation:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        coverImage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, author, isbn, publisher, publicationYear, category, tags, format, totalCopies, availableCopies, createdAt]
...
    Loan:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        bookId:
          type: integer
        book:
          $ref: "#/components/schemas/Book"
        user:
          $ref: "#/components/schemas/User"
        borrowedAt:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
        returnedAt:
          type: string
          format: date-time
          nullable: true
        renewalsCount:
          type: integer
        status:
          type: string
          enum: [ACTIVE, RETURNED, OVERDUE]
      required: [id, userId, bookId, borrowedAt, dueDate, renewalsCount, status]
    BorrowBookBody:
      type: object
      properties:
        bookId:
          type: integer
      required: [bookId]
    PaginatedLoans:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Loan"
        page:
          type: integer
        pageSize:
          type: integer
        totalRecords:
          type: integer
        totalPages:
          type: integer
      required: [data, page, pageSize, totalRecords, totalPages]
...